Change team digest into an intro if its the first one
